# docs/openapi.yaml
openapi: 3.1.0
info:
  title: GoLang User Manager API
  version: 1.0.0
  description: >
    Documentação completa da API do GoLang User Manager (Auth, Users, Profiles, Plans e Subscriptions).
    Compatível com Scalar, Swagger e Postman.
  contact:
    name: Tarik Villalobos
    url: https://github.com/tarikvillalobos

servers:
  - url: http://localhost:8080
    description: Ambiente local

tags:
  - name: Auth
    description: Registro e autenticação de usuários
  - name: Users
    description: CRUD de usuários e gestão de senhas
  - name: Profile
    description: Perfil do usuário autenticado
  - name: Plans
    description: Planos disponíveis
  - name: Subscriptions
    description: Assinaturas dos usuários

paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Registrar novo usuário
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Usuário registrado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Dados inválidos
        '409':
          description: E-mail já cadastrado

  /auth/login:
    post:
      tags: [Auth]
      summary: Login de usuário
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login bem-sucedido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Credenciais inválidas

  /auth/refresh:
    post:
      tags: [Auth]
      summary: Atualizar token JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Novo token JWT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'

  /users:
    get:
      tags: [Users]
      summary: Listar todos os usuários
      responses:
        '200':
          description: Lista de usuários
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'

    post:
      tags: [Users]
      summary: Criar novo usuário
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: Usuário criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

  /users/{id}:
    get:
      tags: [Users]
      summary: Buscar usuário por ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Usuário encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: Usuário não encontrado

    put:
      tags: [Users]
      summary: Atualizar dados de um usuário
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: Usuário atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

    delete:
      tags: [Users]
      summary: Excluir usuário
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Usuário excluído com sucesso

  /users/{id}/password:
    put:
      tags: [Users]
      summary: Atualizar senha do usuário
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePasswordRequest'
      responses:
        '200':
          description: Senha atualizada com sucesso
        '400':
          description: Senha inválida

  /profile:
    get:
      tags: [Profile]
      summary: Obter perfil do usuário autenticado
      responses:
        '200':
          description: Dados do perfil
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'

    put:
      tags: [Profile]
      summary: Atualizar perfil do usuário autenticado
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Perfil atualizado

  /plans:
    get:
      tags: [Plans]
      summary: Listar planos disponíveis
      responses:
        '200':
          description: Lista de planos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Plan'

  /subscriptions:
    post:
      tags: [Subscriptions]
      summary: Criar assinatura de plano
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSubscriptionRequest'
      responses:
        '201':
          description: Assinatura criada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        plan_id:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    Profile:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        phone:
          type: string
        avatar:
          type: string
        bio:
          type: string

    Plan:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: float
        limit_users:
          type: integer

    Subscription:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        plan_id:
          type: integer
        status:
          type: string
        started_at:
          type: string
          format: date-time
        ends_at:
          type: string
          format: date-time

    RegisterRequest:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string

    LoginResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        expires_in:
          type: integer
        user:
          $ref: '#/components/schemas/User'

    RefreshRequest:
      type: object
      properties:
        refresh_token:
          type: string

    CreateUserRequest:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string

    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string

    UpdatePasswordRequest:
      type: object
      required: [old_password, new_password]
      properties:
        old_password:
          type: string
        new_password:
          type: string

    UpdateProfileRequest:
      type: object
      properties:
        phone:
          type: string
        avatar:
          type: string
        bio:
          type: string

    CreateSubscriptionRequest:
      type: object
      required: [user_id, plan_id]
      properties:
        user_id:
          type: integer
        plan_id:
          type: integer